Update mint_nft.cdc
I was having some issues with this transaction and I brought it up in the discord.   The solution I got from the discord was to put the address inside getAccount.  I later tried the change I made below and that worked as well.

let recipient = getAccount(signer.address) in line 22